Internal function to calculate theta parameter

Calculates values of theta matrix within Gibbs sampler. Not for calling directly by users.

get.theta(v, nmaxclust, ntsegm)

Arguments

v

A matrix returned by sample.v

nmaxclust

numeric. A single number indicating the maximum number of clusters to test.

ntsegm

numeric. The total number of time segments from all animal IDs.

Value

A matrix of proportion estimates that represent proportions of different behavioral states per time segment.