Insert gaps for missing behavior proportions in time series

Insert gaps for missing behavior proportions in time series

insert_date_gaps(data, tol, units)

Arguments

data
A data frame containing all of the columns returned by expand_behavior. This includes id, tseg, time1, date, behavior, and prop.

tol
integer. The tolerance (or threshold) of time after which NA breaks should be inserted to facilitate accurate time series plots.

units
character. The units by which tol is specified, and therefore how time differences between dates are measured. These can be "secs", "mins", "hours", "days", or "weeks".

Value

A data frame nearly identical to that originally produced by expand_behavior, but that includes inserted NA values to denote breaks in the date greater than tol. Additionally, two new columns (ymin and ymax) are included within this data frame. These columns can be used within geom_ribbon to produce stacked areas that account for gaps in the time series.

Examples

# \donttest{
#load data
data(tracks.seg)

#select only id, tseg, SL, and TA columns
tracks.seg2<- tracks.seg[,c("id","tseg","SL","TA")]

#summarize data by track segment
obs<- summarize_tsegs(dat = tracks.seg2, nbins = c(5,8))

#cluster data with LDA
res<- cluster_segments(dat = obs, gamma1 = 0.1, alpha = 0.1, ngibbs = 1000,
                       nburn = 500, nmaxclust = 7, ndata.types = 2)

#Extract proportions of behaviors per track segment
theta.estim<- extract_prop(res = res, ngibbs = 1000, nburn = 500, nmaxclust = 7)

#Create augmented matrix by replicating rows (tsegs) according to obs per tseg
theta.estim.long<- expand_behavior(dat = tracks.seg, theta.estim = theta.estim, obs = obs,
                               nbehav = 3, behav.names = c("Encamped","ARS","Transit"),
                               behav.order = c(1,2,3))

#Add gaps when dt > 1 week (in minutes)
theta.estim.list<- df_to_list(dat = theta.estim.long, ind = "id")
theta.estim.long2<- purrr::map(theta.estim.list,
                      ~insert_date_gaps(.x, tol = 60*24*7, units = "min")) %>%
  dplyr::bind_rows()
# }