Segmented tracks for all IDs.

A dataset containing the filtered track data with time segments assigned to all observations on an individual basis.

tracks.seg

Format

A data frame with 14096 rows and 12 variables:

id

ID for each simulated track

date

date, recorded as datetime

x

x coordinate of tracks

y

y coordinate of tracks

step

the step length calculated as the distance between successive locations measured in units

angle

the relative turning angle measured in radians

dt

the time step or sampling interval between datetimes of successive observations

obs

the ordered number of observations per ID before filtering for the primary time step

time1

the ordered number of observations per ID after filtering for the primary time step

SL

discretized step lengths, separated into five bins

TA

discretized turning angles, separated into eight bins

tseg

time segment assigned to a given set of observations per ID